+++ src/App.test.js
@@ -1,9 +1,15 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App, { Welcome } from './App';
 
-test('renders learn react link', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+// test('renders learn react link', () => {
+//   const { getByText } = render(<App />);
+//   const linkElement = getByText(/learn react/i);
+//   expect(linkElement).toBeInTheDocument();
+// });
+
+test('renders welcome form', () => {
+  const { getByText } = Welcome.render();
+  const header = getByText(/Welcome to Household Hunt/i);
+  expect(header).toBeInTheDocument();
 });
